refactor(usePhraseExtraction): extract flow logic into a module
The 370-line hook was mostly pure async orchestration + debug logging. Move
checkExistingExtraction/saveExtractionToDatabase/extractPhrasesFlow into
utils/phraseExtractionFlow.ts (no React deps); the hook is now a 54-line wrapper
that only tracks isExtracting. Trimmed leftover '=== DEBUG ===' console.log
tracing (kept console.error/warn). Typed existingExtraction as PhraseExtraction
(removes a no-explicit-any) and fixed the handleExtraction deps (removes an
exhaustive-deps warning).

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
